+++ src/app/services/content-resources.js
@@ -43,12 +43,11 @@ export default Service.extend({
         return this.get('groupManager').getRecord(id);
       case 'harvesters':
         return this.get('harvesterManager').getRecord(id);
-      case 'uploads':
-        {
-          const oneprovider = this.get('uploadManager.sidebarOneproviders')
-            .findBy('id', id);
-          return oneprovider ? resolve(oneprovider) : reject();
-        }
+      case 'uploads': {
+        const oneprovider = this.get('uploadManager.sidebarOneproviders')
+          .findBy('id', id);
+        return oneprovider ? resolve(oneprovider) : reject();
+      }
       default:
         return reject('No such model type: ' + type);
     }