+++ src/tests/integration/components/membership-visualiser-test.js
@@ -1,22 +1,36 @@
 import { expect } from 'chai';
-import { describe, it, beforeEach } from 'mocha';
+import { describe, it, beforeEach, afterEach } from 'mocha';
 import { setupComponentTest } from 'ember-mocha';
 import hbs from 'htmlbars-inline-precompile';
 import EmberObject, { get, setProperties } from '@ember/object';
 import { registerService, lookupService } from '../../helpers/stub-service';
 import _ from 'lodash';
-import { resolve } from 'rsvp';
+import { resolve, reject } from 'rsvp';
 import Service from '@ember/service';
 import wait from 'ember-test-helpers/wait';
 import $ from 'jquery';
+import TestAdapter from '@ember/test/adapter';
+import Ember from 'ember';
 
 const userEntityId = 'userEntityId';
+
+function groupGri(gropuNo) {
+  return `group.group${gropuNo}EntityId.instance:private`;
+}
+
+function membershipGri(groupNo) {
+  return `group.group${groupNo}EntityId.eff_user_membership,${userEntityId}:private`;
+}
 
 const StoreStub = Service.extend({
   groups: Object.freeze({}),
   memberships: Object.freeze({}),
 
   findRecord(modelName, id) {
+    // Testing edge case when cannot fetch full membership path
+    if (id == membershipGri(2)) {
+      return reject({ id: 'forbidden' });
+    }
     return resolve(this.peekRecord(modelName, id));
   },
 
@@ -36,7 +50,7 @@ function generateNGroups(context, n) {
   const membershipsMap = {};
   const groups = _.times(n, index => {
     const group = EmberObject.create({
-      gri: `group.group${index}EntityId.instance:private`,
+      gri: groupGri(index),
       entityType: 'group',
       name: `groups${index}`,
     });
@@ -45,7 +59,7 @@ function generateNGroups(context, n) {
   });
   groups.forEach((group, index) => {
     const membership = EmberObject.create({
-      gri: `group.group${index}EntityId.eff_user_membership,${userEntityId}:private`,
+      gri: membershipGri(index),
       directMembership: true,
       intermediaries: groups.filter(g => g !== group).mapBy('gri'),
     });
@@ -75,6 +89,20 @@ describe('Integration | Component | membership visualiser', function () {
     });
     generateNGroups(this, 3);
     this.set('user', user);
+
+    // Hack: changing default error handling in tests, due to the Ember bug:
+    // rejected, not caught promises with some error inside (not empty) marks
+    // tests as failed. So e.g. `reject({ id: 'forbidden' })` in store.findRecord,
+    // crashes tests even if it's a correct behaviour of findRecord.
+    this.originalLoggerError = Ember.Logger.error;
+    this.originalTestAdapterException = TestAdapter.exception;
+    Ember.Logger.error = function() {};
+    Ember.Test.adapter.exception = function() {};
+  });
+
+  afterEach(function () {
+    Ember.Logger.error = this.originalLoggerError;
+    Ember.Test.adapter.exception = this.originalTestAdapterException;
   });
 
   it('renders all possible paths', function () {
@@ -82,7 +110,7 @@ describe('Integration | Component | membership visualiser', function () {
       contextRecord=user
       targetRecord=groups.[0]}}`);
     return wait().then(() => {
-      expect(this.$('.membership')).to.have.length(5);
+      expect(this.$('.membership')).to.have.length(4);
     });
   });
 
@@ -108,6 +136,16 @@ describe('Integration | Component | membership visualiser', function () {
         expect(blocksNumber).to.be.gte(prevBlocksNumber);
         prevBlocksNumber = blocksNumber;
       });
+    });
+  });
+
+  it('renders all possible paths when visibleBlocks equals 2', function () {
+    this.render(hbs `{{membership-visualiser
+      contextRecord=user
+      visibleBlocks=2
+      targetRecord=groups.[0]}}`);
+    return wait().then(() => {
+      expect(this.$('.membership')).to.have.length(4);
     });
   });
 });